Many bug fixes (#83)
* Update compiler.py

* debugging

* remove debugging

* num items in batch

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* logs

* Update patching_utils.py

* VLM attention mask

* Update loss_utils.py

* Update loss_utils.py

* Update loss_utils.py

* Update loss_utils.py

* Update loss_utils.py

* Update loss_utils.py

* Recheck

* Update compiler.py

* Update patching_utils.py

* Update patching_utils.py

* Update patching_utils.py

* Update patching_utils.py

* Update compiler.py

* Update patching_utils.py

* suppress errors

* Update compiler.py

* Update patching_utils.py

* Update compiler.py

* Update patching_utils.py

* Update patching_utils.py

* Update patching_utils.py

* Update peft_utils.py

* Update compiler.py

* Update loss_utils.py

* Update loss_utils.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* bug fixes

* Update compiler.py

* Update compiler.py

* Update vision_utils.py

* Update compiler.py

* Update loss_utils.py

* Update loss_utils.py

* Update loss_utils.py

* Update loss_utils.py

* Bug fixes

* Update dataset_utils.py

* Update dataset_utils.py

* Update dataset_utils.py

* Update dataset_utils.py

* Update dataset_utils.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update loss_utils.py

* Update loss_utils.py

* gpu_memory_utilization

* Update temporary_patches.py

* Update vision_utils.py

* Update vision_utils.py

* Update vision_utils.py

* Update vision_utils.py

* Update vision_utils.py

* Update vision_utils.py

* Update vision_utils.py

* Update vision_utils.py

* train on completions VLMs

* Update dataset_utils.py

* Update dataset_utils.py

* Update dataset_utils.py

* Update dataset_utils.py

* VLM train only on completions

* Update loss_utils.py

* Update dataset_utils.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update saving_utils.py

* Update llama_cpp.py

* Update llama_cpp.py

* Update saving_utils.py

* Update saving_utils.py

* Update __init__.py

* Update compiler.py

* Update loss_utils.py

* Update compiler.py

* Update loss_utils.py

* Update loss_utils.py

* Update llama_cpp.py

* Update loss_utils.py

* Update compiler.py

* Update llama_cpp.py

* Update compiler.py

* Update vllm_utils.py

* Update rl_replacements.py

* Update rl_replacements.py

* Update rl_replacements.py

* Update rl_replacements.py

* Update rl_replacements.py

* Update rl_replacements.py

* Update training_utils.py

* Update dataset_utils.py

* Update dataset_utils.py

* Revert "Update dataset_utils.py"

This reverts commit 3b690ad5286e22682f137374d84384e569d20ce7.

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update compiler.py

* Update compiler.py

* Remove prints

* Update compiler.py

* Update saving_utils.py

* Update temporary_patches.py

* Update __init__.py

* Update pyproject.toml

* Update vllm_utils.py

* bug fix #2008 unsloth issue - load_in_4bit = True + fast_inference = True (#79)

* bug fix #2008 unsloth

* non-quant dtype fix

* Update vllm_utils.py

---------


* Update dataset_utils.py

* Update compiler.py

* Update temporary_patches.py

* Gemma 3 fixes

* Update temporary_patches.py

* Update compiler.py

* Update compiler.py

* Gemma 3 fixes

* Update patching_utils.py

* Update compiler.py

* Update compiler.py

* Update patching_utils.py

* Update temporary_patches.py

* Update compiler.py

* Update compiler.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* Update compiler.py

* compiler

* Update gradient_checkpointing.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* causal mask dtype

* Fix checkpoint and save from local file (#74)

* Enhance gradient checkpointing and add original model ID retrieval in saving utilities

* In case adapter_config.json as well

* Update patching_utils.py

* Update patching_utils.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update compiler.py

* Update loss_utils.py

* Update compiler.py

* Update vllm_utils.py

* Update compiler.py

* Update peft_utils.py

* Update rl_replacements.py

* Update vllm_utils.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

* Update temporary_patches.py

---------

+++ unsloth_zoo/loss_utils.py
@@ -168,7 +168,7 @@ def fused_linear_cross_entropy(
     reduction = "sum" if num_items_in_batch is not None else "mean"
     if logit_softcapping == 0: logit_softcapping = None
     loss = linear_cross_entropy(
-        hidden_states,
+        hidden_states.to(lm_weight.dtype),
         lm_weight,
         targets      = labels,
         ignore_index = ignore_index,
